+++ src/main/java/nextstep/console/controller/ConsoleReservationController.java
@@ -4,9 +4,9 @@ import nextstep.console.view.ConsoleReservationView;
 import nextstep.domain.repository.ReservationRepository;
 import nextstep.domain.repository.ThemeRepository;
 import nextstep.dto.CreateReservationRequest;
-import nextstep.repository.JdbcReservationRepository;
-import nextstep.repository.JdbcThemeRepository;
-import nextstep.service.ReservationService;
+import nextstep.web.repository.JdbcReservationRepository;
+import nextstep.web.repository.JdbcThemeRepository;
+import nextstep.web.service.ReservationService;
 import nextstep.utils.DataSourceUtil;
 import org.springframework.jdbc.core.JdbcTemplate;
