+++ src/main/java/nextstep/web/repository/JdbcThemeRepository.java
@@ -1,4 +1,4 @@
-package nextstep.repository;
+package nextstep.web.repository;
 
 import nextstep.domain.Theme;
 import nextstep.domain.repository.ThemeRepository;
@@ -9,8 +9,6 @@ import org.springframework.jdbc.core.namedparam.BeanPropertySqlParameterSource;
 import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 import org.springframework.stereotype.Repository;
 
-import java.sql.Date;
-import java.sql.Time;
 import java.util.List;
 import java.util.Optional;
